refactor: dry up page scanning
