+++ src/lib/syncPagePath.ts
@@ -1,10 +1,6 @@
-import type { PageMetadata } from './readPage.js';
 import type { GuidesRequestRepresentation, ProjectRepresentation } from './types/index.js';
 import type DocsMigrateCommand from '../commands/docs/migrate.js';
 import type DocsUploadCommand from '../commands/docs/upload.js';
-
-import fs from 'node:fs/promises';
-import path from 'node:path';
 
 import chalk from 'chalk';
 import ora from 'ora';
@@ -15,9 +11,8 @@ import { fix, writeFixes } from './frontmatter.js';
 import isCI from './isCI.js';
 import { oraOptions } from './logger.js';
 import promptTerminal from './promptWrapper.js';
-import readdirRecursive from './readdirRecursive.js';
 import { fetchMappings, fetchSchema } from './readmeAPIFetch.js';
-import readPage from './readPage.js';
+import { findPages, type PageMetadata } from './readPage.js';
 import { categoryUriRegexPattern, parentUriRegexPattern } from './types/index.js';
 
 /**
@@ -200,15 +195,6 @@ export default async function syncPagePath(this: CommandsThatSyncMarkdown) {
   const { path: pathInput }: { path: string } = this.args;
   const { key, 'dry-run': dryRun, 'skip-validation': skipValidation } = this.flags;
 
-  const allowedFileExtensions = ['.markdown', '.md', '.mdx'];
-
-  const stat = await fs.stat(pathInput).catch(err => {
-    if (err.code === 'ENOENT') {
-      throw new Error("Oops! We couldn't locate a file or directory at the path you provided.");
-    }
-    throw err;
-  });
-
   // check whether or not the project has bidirection syncing enabled
   // if it is, validations must be skipped to prevent frontmatter from being overwritten
   const headers = new Headers({ authorization: `Bearer ${key}` });
@@ -236,41 +222,7 @@ export default async function syncPagePath(this: CommandsThatSyncMarkdown) {
     }
   }
 
-  let files: string[];
-
-  if (stat.isDirectory()) {
-    const fileScanningSpinner = ora({ ...oraOptions() }).start(
-      `🔍 Looking for Markdown files in the \`${pathInput}\` directory...`,
-    );
-    // Filter out any files that don't match allowedFileExtensions
-    files = readdirRecursive(pathInput).filter(file =>
-      allowedFileExtensions.includes(path.extname(file).toLowerCase()),
-    );
-
-    if (!files.length) {
-      fileScanningSpinner.fail(`${fileScanningSpinner.text} no files found.`);
-      throw new Error(
-        `The directory you provided (${pathInput}) doesn't contain any of the following file extensions: ${allowedFileExtensions.join(
-          ', ',
-        )}.`,
-      );
-    }
-
-    fileScanningSpinner.succeed(`${fileScanningSpinner.text} ${files.length} file(s) found!`);
-  } else {
-    const fileExtension = path.extname(pathInput).toLowerCase();
-    if (!allowedFileExtensions.includes(fileExtension)) {
-      throw new Error(
-        `Invalid file extension (${fileExtension}). Must be one of the following: ${allowedFileExtensions.join(', ')}`,
-      );
-    }
-
-    files = [pathInput];
-  }
-
-  this.debug(`number of files: ${files.length}`);
-
-  let unsortedFiles = files.map(file => readPage.call(this, file));
+  let unsortedFiles = await findPages.call(this, pathInput);
 
   if (!skipValidation) {
     const validationSpinner = ora({ ...oraOptions() }).start('🔬 Validating frontmatter data...');
